requirements.txt
streamlit==1.18.1
pandas==1.5.3
numpy==1.24.3
opencv-python==4.7.0.72
pillow==9.5.0
scikit-learn==1.2.2
xgboost==1.7.6
shap==0.41.0
vaderSentiment==3.3.2